inventory is updating according to quantity of orderItem has been added

+++ Repositories/Implementations/SqlInventoryRepository.cs
@@ -1,4 +1,5 @@
 using FulfillmentCenter.Data;
+using FulfillmentCenter.DTOs.Requests;
 using FulfillmentCenter.Entities;
 using FulfillmentCenter.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
@@ -8,14 +9,14 @@ namespace FulfillmentCenter.Repositories.Implementations;
 public class SqlInventoryRepository : IInventoryRepository
 {
     private FulfillmentCenDbContext _context;
-    //public List<Inventory> Inventories;
-    //private bool isCached;
+    public List<Inventory> Inventories;
+    private bool _isCached;
     
     public SqlInventoryRepository(FulfillmentCenDbContext context)
     {
         _context = context;
-        /*Inventories = Read().Result;
-        isCached = true;*/
+        Inventories = Read().Result;
+        _isCached = true;
     }
 
     public async Task Create(Inventory inventory)
@@ -24,6 +25,7 @@ public class SqlInventoryRepository : IInventoryRepository
         {
             await _context.Inventory.AddAsync(inventory);
             await _context.SaveChangesAsync();
+            _isCached = false;
         }
         catch (Exception e)
         {
@@ -32,42 +34,53 @@ public class SqlInventoryRepository : IInventoryRepository
         }
     }
 
-    public async void Delete(Guid id)
+    public async Task Delete(Guid id)
     {
         var inventoryToDelete = await _context.Inventory.FirstOrDefaultAsync(inventory => inventory.Id == id);
+        if(inventoryToDelete == null)
+        {
+            throw new ArgumentNullException();
+        }
         _context.Inventory.Remove(inventoryToDelete);
         await _context.SaveChangesAsync();
     }
 
     public async Task<List<Inventory>> Read()
-    {
+    {//All Read() methods load the entire table into memory as a List<T>. No filtering, no Where, no pagination. This will not scale
         List<Inventory> inventories;
-        //if (isCached == false)
-        //{
-        try
+        if (_isCached == false)
         {
-            inventories = await _context.Inventory.ToListAsync();
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            throw;
-        }
-            //isCached = true;
+            try
+            {
+                inventories = await _context.Inventory.ToListAsync();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
+            _isCached = true;
             return inventories;
-        //}
+        }
 
-        //return Inventories;
+        return Inventories;
     }
+    
     public async Task UpdateInventory(Inventory inventory)
     {
         try
         {
             var inventoryToUpdate = await _context.Inventory.FirstOrDefaultAsync(inv =>
                 inv.ProductId == inventory.ProductId && inv.DistributionCenterId == inventory.DistributionCenterId);
-            inventoryToUpdate.Quantity += 1;
+
+            if (inventoryToUpdate == null)
+            {
+                throw new ArgumentNullException();
+            }
+            inventoryToUpdate.Quantity = inventory.Quantity;
 
             await _context.SaveChangesAsync();
+            _isCached = false;
         }
         catch (Exception e)
         {
@@ -82,8 +95,27 @@ public class SqlInventoryRepository : IInventoryRepository
             });*/
     }
 
-    public async Task UpdateInventoryQuantity(Inventory inventory)
+    public async Task UpdateInventoryQuantity(UpdateInventoryDto inventory)
     {
-        _context.SaveChangesAsync();
+        try
+        {
+            var inventoryToUpdate = await _context.Inventory.FirstOrDefaultAsync(inv =>
+                inv.ProductId == inventory.ProductId);
+
+            if (inventoryToUpdate == null)
+            {
+                throw new ArgumentNullException();
+            }
+            
+            inventoryToUpdate.Quantity = inventory.Quantity;
+
+            await _context.SaveChangesAsync();
+            _isCached = false;
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e);
+            throw;
+        }
     }
 }
